[FEAT]: Fix Rename Service Tpye to Service Category
