+++ src/pages/MedicalService/ServiceCategories/ServiceCategoriesPage.module.scss
