+++ src/pages/MedicalService/ServiceCategories/components/ServiceCategoryCard.module.scss
