--- src/pages/MedicalService/ServiceTypes/index.ts
@@ -1,1 +1,0 @@
-export { default } from './ServiceTypesPage';
